+++ pdf/lib/widgets.dart
@@ -42,3 +42,4 @@ part 'widgets/table.dart';
 part 'widgets/text.dart';
 part 'widgets/theme.dart';
 part 'widgets/widget.dart';
+part 'widgets/wrap.dart';