+++ pdf/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add better debugPaint on Align Widget
 * Fix Transform placement when Alignment and Origin are Null
 * Add Transform.rotateBox constructor
+* Add Wrap Widget
 
 ## 1.3.15